5March2024/requirements.txt
streamlit
streamlit_option_menu
pandas
numpy
plotly.express
datetime
altair